+++ Magic/src/main/java/com/elmakers/mine/bukkit/action/builtin/AllEntitiesAction.java
@@ -46,7 +46,6 @@ public class AllEntitiesAction extends CompoundEntityAction
     public void addEntities(CastContext context, List<WeakReference<Entity>> entities)
     {
         Spell spell = context.getSpell();
-        Entity sourceEntity = context.getMage().getEntity();
         Location sourceLocation = context.getLocation();
 
         if (sourceLocation == null && !targetAllWorlds)