+++ MagicAPI/src/main/java/com/elmakers/mine/bukkit/api/block/UndoList.java
@@ -11,11 +11,13 @@ import org.bukkit.World;
 import org.bukkit.block.Block;
 import org.bukkit.entity.Entity;
 import org.bukkit.entity.EntityType;
+import org.bukkit.potion.PotionEffectType;
 
 import com.elmakers.mine.bukkit.api.action.CastContext;
 import com.elmakers.mine.bukkit.api.batch.Batch;
 import com.elmakers.mine.bukkit.api.entity.EntityData;
 import com.elmakers.mine.bukkit.api.magic.Mage;
+import com.elmakers.mine.bukkit.api.magic.MageModifier;
 import com.elmakers.mine.bukkit.api.spell.Spell;
 
 public interface UndoList extends BlockList, Comparable<UndoList> {
@@ -62,6 +64,10 @@ public interface UndoList extends BlockList, Comparable<UndoList> {
     void move(Entity entity);
     void modifyVelocity(Entity entity);
     void addPotionEffects(Entity entity);
+    void addPotionEffectForRemoval(Entity entity, PotionEffectType potionEffectType);
+
+    void addModifier(Entity entity, MageModifier modifiers);
+    void addModifierForRemoval(Entity entity, String modifierKey);
 
     void convert(Entity entity, Block block);
     void fall(Entity entity, Block block);